modify according to comments

+++ src/Storages/MergeTree/MergeTreeDataPartCNCH.cpp
@@ -628,8 +628,6 @@ MergeTreeDataPartChecksums::FileChecksums MergeTreeDataPartCNCH::loadPartDataFoo
 {
     const String data_file_path = fs::path(getFullRelativePath()) / DATA_FILE;
     size_t data_file_size = volume->getDisk()->getFileSize(data_file_path);
-    if (!volume->getDisk()->fileExists(data_file_path))
-        throw Exception(ErrorCodes::NO_FILE_IN_DATA_PART, "No data file of part {} under path {}", name, data_file_path);
 
     auto data_file = openForReading(volume->getDisk(), data_file_path, MERGE_TREE_STORAGE_CNCH_DATA_FOOTER_SIZE);
 
@@ -1135,7 +1133,7 @@ void MergeTreeDataPartCNCH::removeImpl(bool keep_shared_data) const
     try
     {
         disk->removeFile(path_on_disk / "data");
-        disk->removeDirectory(path_on_disk);
+        if (disk->getType() != DiskType::Type::ByteS3) disk->removeDirectory(path_on_disk);
     }
     catch (...)
     {
@@ -1311,14 +1309,6 @@ void MergeTreeDataPartCNCH::preload(UInt64 preload_level, ThreadPool & pool, UIn
     }
 
     pool.scheduleOrThrowOnError([this, part_path, full_path, level = preload_level, segments = std::move(segments), cb = std::move(callback), disk_cache = cache] {
-        if (!volume->getDisk()->fileExists(part_path))
-        {
-            LOG_WARNING(storage.log, "Can't find {} when preload level: {} on caching", full_path + DATA_FILE, level);
-            if (cb)
-                cb("Can't find part file!", 0);
-            return;
-        }
-
         String last_exception{};
         int real_cache_segments_count = 0;
